Update fire thrower script and move it to the script folder
