+++ Assets/Scripts/Interactables/Traps.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22664e1a96e7e444aa8eced00ab0cb7d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
