+++ Assets/Scripts/Interactables/Traps/FireThrower.cs.meta
